+++ src/routes/Calendar/List/index.ts
@@ -1,0 +1,5 @@
+// Copyright (C) 2017-2024 Smart code 203358507
+
+import List from './List';
+
+export default List;
